refactor: modularize registry command structure
Split monolithic registry.rs into organized submodules for better maintainability:
- formatters.rs: Display and output formatting
- helpers.rs: Common utilities and configuration
- marketplace.rs: Marketplace browsing functionality
- repo_ops.rs: Repository management operations
- skill_ops.rs: Skill-related operations

Also includes:
- Add GitHub Actions workflow for auto-deployment
- Add comprehensive test runner script
- Update dependencies
