+++ .gitignore
@@ -321,3 +321,4 @@ ORIG_HEAD
 commondir
 gitdir
 index
+.newton
